DeepTrader/requirements.txt
snowflake-connector-python==3.7.0
pandas==2.0.3
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
ta==0.10.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
scikit-learn==1.3.0
tqdm==4.66.1
dash==2.14.1
dash-bootstrap-components==1.5.0 